Wizard/Packages: Move re-fetching logic where it belongs
The re-fetching logic applies only to packages rendered in the package table. It was probably left in the search during the migration, but since search fetches fresh results every time, it doesn't need any re-fetching in edit mode.
